Move the discovery service manager code under cmd/, like the rest of commands
